Rename .es.js files to just .js
