+++ gradle/apps/npm/billboardjs-npm-portlet/src/main/resources/META-INF/resources/lib/index.js
