+++ gradle/apps/npm/isomorphic-npm-portlet/src/main/resources/META-INF/resources/lib/index.js
