+++ gradle/apps/npm/jquery-npm-portlet/src/main/resources/META-INF/resources/lib/index.js
