+++ gradle/apps/npm/metaljs-npm-portlet/package.json
@@ -13,7 +13,7 @@
 		"liferay-npm-bundler-preset-standard": "1.2.2",
 		"metal-tools-soy": "4.2.7"
 	},
-	"main": "lib/index.es.js",
+	"main": "lib/index.js",
 	"name": "metaljs-npm-portlet",
 	"private": true,
 	"scripts": {
