+++ gradle/apps/npm/metaljs-npm-portlet/src/main/resources/META-INF/resources/lib/CloseHeader.js
