+++ gradle/apps/npm/react-npm-portlet/package.json
@@ -12,7 +12,7 @@
 		"liferay-npm-bundler": "1.2.2",
 		"liferay-npm-bundler-preset-react": "1.2.2"
 	},
-	"main": "lib/index.es.js",
+	"main": "lib/index.js",
 	"name": "react-npm-portlet",
 	"private": true,
 	"scripts": {
