+++ gradle/apps/npm/react-npm-portlet/src/main/resources/META-INF/resources/lib/index.js
