+++ gradle/apps/npm/simple-npm-portlet/src/main/resources/META-INF/resources/lib/index.js
