+++ gradle/apps/npm/vuejs-npm-portlet/src/main/resources/META-INF/resources/lib/index.js
